% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleioplot.R
\name{pleioplot}
\alias{pleioplot}
\title{pleioplot}
\usage{
pleioplot(
  factors,
  ldf,
  gwas_beta,
  eta,
  eta_ci,
  lmp_input,
  eta_col = c("#E77E23", "#309F86"),
  link_col = c("#F51929", "#2341F6"),
  link_hex = 3,
  size_scale = 0.6,
  study_palette = default_palette
)
}
\arguments{
\item{factors}{example}
}
\value{
A matrix of the infile
}
\description{
This function generate pleioplot.
}
