% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleioplot.R
\name{reorder_using_aoe}
\alias{reorder_using_aoe}
\title{Reorder the variables using the angular order of the eigenvectors.}
\usage{
reorder_using_aoe(corr)
}
\description{
Reorder the variables using the angular order of the eigenvectors.
}
