% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{reorder_using_fpc}
\alias{reorder_using_fpc}
\title{Reorder the variables using the first principal component.}
\usage{
reorder_using_fpc(corr)
}
\description{
Reorder the variables using the first principal component.
}
