% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleioplot.R
\name{reorder_using_hclust}
\alias{reorder_using_hclust}
\title{Reorder the variables using hclust (Hierarchical Clustering).}
\usage{
reorder_using_hclust(corr, hclust.method)
}
\description{
Reorder the variables using hclust (Hierarchical Clustering).
}
