% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleioplot.R
\name{pleioplot}
\alias{pleioplot}
\title{pleioplot}
\usage{
pleioplot(
  snp,
  traits,
  rg_matrix,
  sumstats,
  pleioin,
  pleiores,
  h2,
  snp_reference,
  eta_col = c("#E77E23", "#309F86"),
  rg_col = c("#F51929", "#2341F6"),
  link_hex = 3,
  size_scale = 0.6,
  pleioplot_palette = paste(rainbow_palete_hex, default_alpha, sep = "")
)
}
\arguments{
\item{factors}{example}
}
\value{
A matrix of the infile
}
\description{
This function can generate circo plot.
}
